requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
